+++ changelog.md
@@ -13,6 +13,14 @@ will not have contiguous patch numbers. Initial major and minor releases will be
 in this file without a patch number. Patch version will be included for bug fix releases, but
 may not exactly match a publicly released version.
 
+## Unreleased
+
+### Added
+
+#### NeoExpress
+
+* Added `rpc.MaxIteratorResultItems` setting
+
 ## [3.1.38] 2021-14
 
 ### Added